Testes unitários sendo implementados e modificações no banco de dados

+++ Mercado.MVC/Models/VendaModel.cs
@@ -10,7 +10,7 @@ namespace Mercado.MVC.Models
         [Key]
         public int Id { get; set; }
         [Required(ErrorMessage = "O Campo é obrigatório!")]
-        public int Quantidade { get; set; }
+        public double Quantidade { get; set; }
         [Required(ErrorMessage = "O Campo é obrigatório!")]
         [ForeignKey("Produto")]
         public int IdProduto { get; set; }
